Two apps that work together
