+++ query_app/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "multimodal-CLIP-demo"
-version = "2026.2.8"
+name = "multimodal-CLIP-demo__query_app"
+version = "2026.3.13"
 readme = "README.md"
 license = "Apache-2.0"
 
@@ -8,19 +8,10 @@ dependencies = [
   "typing",
   "fastapi[standard]",
   "jinja2",
-  "pillow",
   "psycopg[binary]",
   "python-dotenv",
-  "torch",
-  "huggingface_hub",
-  #"transformers",
-  # Version 5 of the HuggingFace transformers package introduced a
-  # breaking change to the output of get_image_features. For now
-  # stick with an older version.
-  "transformers==4.57.3",  # try an old version of transformers
-
 ]
 
 [tool.setuptools.packages]
-# We've just got our app.py at the top level, nothing to look for
+# We've just got our clip_app.py at the top level, nothing to look for
 find = {namespaces = false}
